More details on the conversion steps.
Really need to record the steps I'm taking. Already rolled
some changes back so clearly this is a mistake-prone process.

So this will live in a "conversion" branch which I'll eventually
use to make a "master" branch (staging each commit with the proper
author attributes and the best commit message I can derive - later
dirs seem to have announcement files so that helps).